cleanup, default db id config
