+++ discord_bot/terminal_cmds.py
@@ -260,7 +260,7 @@ async def set_bot_presence(bot: "Bot") -> None:
                 return bot.log.warning("Failed to set update flag.")
 
             try:
-                await bot.change_presence(activity=discord.Game(name=new_presence))
+                await bot.change_presence(activity=discord.Activity(type=discord.ActivityType.watching, name=new_presence))
                 bot.log.info(
                     "Config change, presence: {} -> {}".format(
                         config.get("presence"), new_presence
@@ -437,6 +437,7 @@ def show_help(bot: "Bot") -> None:
         "wipebot": 'Wipes the bot"s configuration files.',
         "aliases": "Lists all command aliases.",
         "debug": "Toggles debug mode.",
+        "setdb": "Sets the default DB ID for AskDB.",
     }
 
     try:
@@ -496,6 +497,7 @@ def show_aliases(bot: "Bot") -> None:
         "wipebot": ["wipeconfig", "wipe", "wb"],
         "alias": ["aliases", "a"],
         "debug": ["d"],
+        "setdb": ["defaultdb", "dbd", "dbid"],
     }
 
     try:
@@ -536,3 +538,26 @@ def ping(bot: "Bot") -> None:
         bot.log.info("Pong!")
     except Exception as e:
         bot.log.error(f"Error in ping function: {e}")
+        
+
+def set_default_db_id(bot: "Bot") -> None:
+    """_summary_
+
+    Args:
+        bot (Bot): _description_
+    """
+    bot.log.info("Current default DB ID: {}".format(bot.default_db_id))
+    change_default_db_id = get_boolean_input(bot, "Would you like to change the default DB ID? (y/n) ")
+            
+    try:
+        if change_default_db_id == True:
+            new_db_id = input("Enter new default DB ID: ")
+            bot.default_db_id = new_db_id
+            new_data = {"default_db_id": new_db_id}
+            update_config(bot.config_file, new_data)
+            bot.log.info("Default DB ID changed.")
+            
+        else:
+            bot.log.info("Default DB ID not changed.")
+    except Exception as e:
+        bot.log.error(f"Error in set_default_db_id function: {e}")
